arranged code
