+++ config.py
@@ -1,3 +1,4 @@
 ADDRESS = "localhost"
 PORT = 5001
 INPUT_CIRCUIT_FILENAME = 'input_circuit.json'
+RSA_bits = 1024
