--- ot/mulinv.py
@@ -1,19 +1,0 @@
-
-def egcd(b, n):
-    """
-    Takes positive integers a, b as input, and return a triple (g, x, y), such that 
-    ax + by = g = gcd(a, b).
-    """
-    x0, x1, y0, y1 = 1, 0, 0, 1
-    while n != 0:
-        q, b, n = b // n, n, b % n
-        x0, x1 = x1, x0 - q * x1
-        y0, y1 = y1, y0 - q * y1
-    return b, x0, y0
-
-
-# x = mulinv(b) mod n, (x * b) % n == 1
-def mulinv(b, n):
-    g, x, _ = egcd(b, n)
-    if g == 1:
-        return x % n
