+++ utils/crypto.py
@@ -1,4 +1,9 @@
+from hashlib import sha256
+from random import SystemRandom
+
 from cryptography.fernet import Fernet
+
+crypto_rand = SystemRandom()
 
 
 def encrypt(data, key):
@@ -15,5 +20,9 @@ def get_label():
     return Fernet.generate_key()
 
 
-def keys_to_int(x):
-    return {int(k): v for k, v in x}
+def hasher(b):
+    return sha256(b).hexdigest()
+
+
+def rand_int(n):
+    return crypto_rand.randrange(n)
